Extend supported containers, fix some callback types, and improve container support (#122)
* clang2il.go: Store additional type information

* clangexec.go: Modernize

* intermediate.go: Simplify container detections, extend map support, and add helper methods

* config-allowlist.go: Unblock additional container types

* emitcabi.go: Improve cleanup and fix some callback types

* emitzig.go: Improve hash map container support

* README: Update projection information

* Update bindings and wrappers

+++ src/extras-karchive/libkcompressiondevice.cpp
@@ -91,9 +91,9 @@ void KCompressionDevice_SetSkipHeaders(KCompressionDevice* self) {
 bool KCompressionDevice_Seek(KCompressionDevice* self, long long param1) {
     auto* vkcompressiondevice = dynamic_cast<VirtualKCompressionDevice*>(self);
     if (vkcompressiondevice && vkcompressiondevice->isVirtualKCompressionDevice) {
-        return self->seek(static_cast<long long>(param1));
+        return self->seek(static_cast<qint64>(param1));
     } else {
-        return ((VirtualKCompressionDevice*)self)->seek(static_cast<long long>(param1));
+        return ((VirtualKCompressionDevice*)self)->seek(static_cast<qint64>(param1));
     }
 }
 
@@ -235,9 +235,9 @@ bool KCompressionDevice_QBaseSeek(KCompressionDevice* self, long long param1) {
     auto* vkcompressiondevice = dynamic_cast<VirtualKCompressionDevice*>(self);
     if (vkcompressiondevice && vkcompressiondevice->isVirtualKCompressionDevice) {
         vkcompressiondevice->setKCompressionDevice_Seek_IsBase(true);
-        return vkcompressiondevice->seek(static_cast<long long>(param1));
+        return vkcompressiondevice->seek(static_cast<qint64>(param1));
     } else {
-        return self->KCompressionDevice::seek(static_cast<long long>(param1));
+        return self->KCompressionDevice::seek(static_cast<qint64>(param1));
     }
 }
